prototype - Styled render callback and withStyles refactoring
